Removing update to Publisher title
https://github.com/theodi/publisher/issues/135

When Panopticon artefact changes, Publisher edition triggers update_from_artefact, which updates
the title, amongst other things.

Removing the update to title. This action is triggered when there
is an update to the artefact in Panopticon. Agreed with @pezolio
on 17 June to comment out, in case VERY BAD THINGS happen.
slug is the prime linkage, but think that section, department and
business_proposition are mastered in Panopticon / artefact, so
should maintain them there, and ensure that existing editions are
updated.